Fixed: eBay customers (with names and postcodes) will not be duplicated
When new eBay customers with names and postcodes are added, the
is searched for matching postcodes and names. If a matching customer is
found, their ebay_user_id is set instead of adding a new customer.

Currently uses an exact name match so will not work too well in
practice. Should be improved with more 'fuzzy' name matching.